third_party/WebKit/LayoutTests/editing/deleting/delete-first-list-item.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<ul>',
                '<li style="font-weight: bold">|</li>',
                '<li>bar</li>',
             '</ul>',
        '</div>',
    ].join(''),
    'delete',
    [
        '<div contenteditable>',
            '<div><b>|<br></b></div>',
            '<ul>',
                '<li>bar</li>',
             '</ul>',
        '</div>',
    ].join('')),
    'Delete from empty list item with keeping style');
</script>
